Rename config package to app package
